refactor: don't enforce nullability on refernce values
To make the Reference class more versatile, let's drop the implied
nullability of reference types, allowing the developer to decide if the
value should be nullable or not. Also expose a shallower, read-only
reference interface to make it more explicit to anyone reading code
using references that this reference will not be modified.