feat: introduce a bottom sheet dimensions listenable
This replaces the dragHandleHeight and minContentHeight references with
a unified container for those values which is also a listenable which
notifies its listener whenever the bottom sheet layout is marked dirty.
This will be useful for creating dependencies on the dimensions of the
bottom sheet. For example, let's say we have a scaffold and we display
our persistent bottom sheet in an overlay above the scaffold. To make
sure our bottom sheet does not cover up content in scaffold's body, we
will insert a placeholder into scaffold's bottom navigation bar spot.
This placeholder will use the aforementioned dimensions to size itself.
But now imagine a scenario when there is some ongoing animation that
makes the drag handle of the bottom sheet change its dimensions. The
placeholder will stay the same size, despite the dimensions changing,
because the bottom sheet is in a different subtree, so the placeholder
is not marked dirty for layout. Now, with the listenable API, which our
dimensions provide, the placeholder can listen for notifications and
mark itself dirty whenever a notification arrives.