feat: implement a helper widget for overlaying a sheet over a scaffold
Persistent bottom sheets are typically always visible, even when they
are in their "closed" position, there is still a drag handle sticking
out to allow users to open them up again. This poses a problem for the
Scaffold which is displayed below the bottom sheet, because the drag
handle covers up some of the content in the scaffold's body. This new
helper widget provides a solution for this scenario by offering a
builder method which provides a placeholder widget which can be just
plugged into the scaffold's bottom navigation bar slot. This placeholder
will automatically mark itself dirty whenever it receives a notification
from the bottom sheet that it was marked dirty, which makes sure that
the size of the placeholder is kept in sync with the size of the drag
handle, even if it contains some animations which make it change
dimensions dynamically.