feat: add a layout marked dirty callback to the layout/height observer
This adds a callback to the former height observer which is called
whenever its child's layout is marked dirty. This is useful for
notifying the bottom sheet that one of its dependencies changed
dimensions, so that the bottom sheet can also relayout itself. This also
renames the widget from height to layout observer, as it's a more
generic name for observing multiple layout properties, not only height.