refactor: use CurvedAnimation instead of custom Reference class
It turns out the Flutter framework already provides a class which
gives us a mutable reference to an animation curve, it's the
CurvedAnimation. On top of that, this class is also an animation itself,
resulting in some nice code simplification. This also removes the
reference class, as it is now unused.