feat: add an API to allow dependencies to mark the bottom sheet dirty
For example, if the bottom sheet depends on the size of a navigation
bar, then whenever that navigation bar changes its dimensions (e.g. due
to an animation), the bottom sheet must also change its dimensions.
Typically, the dependency will be located in another subtree, which
may not propagate a relayout to the bottom sheet due to a relayout
boundary on the way. This is why we need an explicit API for the
dependencies to let the bottom sheet know that it should relayout
itself.