refactor: pack all features into a single widget
In testing, it turns out that the current approach of measuring layouts
overlooked a very important rule of layout in the Flutter framework:
widgets are laid out in depth-first order of the tree.

To understand why that is a problem, let's follow an example scenario:

  1. The navigation bar's height is changed in its widget configuration.
  2. The navigation bar is marked dirty for layout.
  3. The PersistentBottomSheetOverlay is indirectly marked dirty for
     layout by the navigation bar's LayoutObserver.
  4. The bottom navigation bar placeholder, passed down to the overlaid
     widget, is also indirectly marked dirty for layout by the
     navigation bar's LayoutObserver.

All of this seems logical: changing the navigation bar's height results
in at least three widgets being marked dirty for layout. The navigation
bar itself obviously, because its widget configuration changed, the
overlay because it might need to reposition its children based on the
new dimensions, and the placeholder because it might need to adjust its
size according to the new dimensions of the navigation bar.

Now, for the next points let's assume that the navigation bar is deeper
in the tree than the bottom navigation bar placeholder, which is not
entirely unrealistic. With this assumption, here's what happens next:

  5. The PersistentBottomSheetOverlay is laid out first, as it's the
     shallowest of the three widgets.
  6. The PersistentBottomSheetOverlay asks its first child, the
     navigation bar, to lay itself out. However, because the navigation
     bar is deeper in the tree, we are likely to encounter an
     intermediate widget, such as an Align widget to position the bar
     at the bottom, rather than the navigation bar itself. Since this
     intermediate widget is not marked dirty for layout, it reuses its
     previous layout.
  7. The overlay then asks the overlaid widget to lay itself out. We
     assume a similar situation occurs here: the previous layout is
     reused. At this point, the overlay completes its layout.
  8. The framework walks the tree in depth-first order looking for
     another widget to layout. That next widget, based on our
     assumption, is the bottom navigation bar placeholder.

And this is the problematic point. The bottom navigation bar placeholder
depends on the result of the navigation bar's layout, but because of the
depth-first order, the navigation bar is going to be laid out after the
placeholder, resulting in the placeholder picking wrong dimensions by
using stale data.

This issue is fixed by significantly simplifying the whole package,
eliminating the need for a bottom navigation bar placeholder altogether.